#97 - switch body tab to own class
